feat(now): migrated now section to now/index.js
